requirements.txt
geographiclib==1.50
geopy==2.1.0
ImageHash==4.2.0
numpy==1.20.1
Pillow==8.1.1
PyWavelets==1.1.1
scipy==1.6.1
six==1.15.0
